requirements.txt
streamlit
plotly.express
pinotdb